+++ ash/app_list/views/apps_grid_view_test_api.h
@@ -49,7 +49,7 @@ class AppsGridViewTestApi {
 
   void PressItemAt(int index);
 
-  size_t TilesPerPage(int page) const;
+  int TilesPerPage(int page) const;
 
   int AppsOnPage(int page) const;
