+++ ash/shelf/shelf_unittest.cc
@@ -71,7 +71,7 @@ class ShelfTest : public AshTestBase {
 // Confirms that ShelfItem reflects the appropriated state.
 TEST_F(ShelfTest, StatusReflection) {
   // Initially we have the app list.
-  size_t button_count = test_api()->GetButtonCount();
+  int button_count = test_api()->GetButtonCount();
 
   // Add a running app.
   ShelfItem item;
@@ -93,7 +93,7 @@ TEST_F(ShelfTest, StatusReflection) {
 // browser test we check this here.
 TEST_F(ShelfTest, CheckHoverAfterMenu) {
   // Initially we have the app list.
-  size_t button_count = test_api()->GetButtonCount();
+  int button_count = test_api()->GetButtonCount();
 
   // Add a running app.
   ShelfItem item;
