+++ ash/system/phonehub/camera_roll_view.cc
@@ -109,7 +109,7 @@ CameraRollView::CameraRollItemsView::~CameraRollItemsView() = default;
 
 void CameraRollView::CameraRollItemsView::AddCameraRollItem(
     views::View* camera_roll_item) {
-  size_t view_size = camera_roll_items_.view_size();
+  int view_size = camera_roll_items_.view_size();
   camera_roll_items_.Add(camera_roll_item, view_size);
   AddChildView(camera_roll_item);
 }
@@ -136,7 +136,7 @@ gfx::Size CameraRollView::CameraRollItemsView::CalculatePreferredSize() const {
 void CameraRollView::CameraRollItemsView::Layout() {
   views::View::Layout();
   CalculateIdealBounds();
-  for (size_t i = 0; i < camera_roll_items_.view_size(); ++i) {
+  for (int i = 0; i < camera_roll_items_.view_size(); ++i) {
     auto* thumbnail = camera_roll_items_.view_at(i);
     thumbnail->SetBoundsRect(camera_roll_items_.ideal_bounds(i));
   }
@@ -159,7 +159,7 @@ gfx::Point CameraRollView::CameraRollItemsView::GetCameraRollItemPosition(
 }
 
 void CameraRollView::CameraRollItemsView::CalculateIdealBounds() {
-  for (size_t i = 0; i < camera_roll_items_.view_size(); ++i) {
+  for (int i = 0; i < camera_roll_items_.view_size(); ++i) {
     gfx::Rect camera_roll_item_bounds =
         gfx::Rect(GetCameraRollItemPosition(i), GetCameraRollItemSize());
     camera_roll_items_.set_ideal_bounds(i, camera_roll_item_bounds);
